Add thread safety
Recording an event/code object and dumping JSON are now thread safe. Only a single
thread can perform any one of these operations at a given time. This is to prevent
array out of bounds exceptions caused by an event or code object being added as
JSON is being written out (and the buffer is copied).