Hijack ServletFilter
If a user is making a request to `/_appmap/record`, we don't want a ServletFilter
blocking that request before it can be handled. As a result, we now transform all
filters and block them them from running. Instead, we immediately call the next
filter in the chain until the request is handled.