Prevent duplicate I/O events
For receiver classes that extend EventProcessorLock,
`onEnterLock`/`OnExitLock` will only be called once for the first
instance of an event with a unique lock key. For example, it's not
uncommon for overloaded methods to call eachother, but this can result
in two or more recorded SQL queries for a single real query.