Make `bin/test` interruptible, interactive
Start the docker container with `-i` so it can be interrupted. If
there's a tty available (e.g. when run by a developer) use it. Pass
arguments, allowing the option of specifying what `entrypoint` should
run.

This allows an interactive invocation of the test environment, for
example

```
$ bin/test /bin/bash
```

which will set up the petclinic server, then start a shell.