Queue events until another event is written
This allows event handlers to modify the previous event. An example may
be writing a normalized path to a previous http_server_request. This
method is somewhat naive and not very flexible, falling apart if another
unintended method is recorded between two events that should be linked.
In the future, it'd be better to walk the events of the call stack
and/or introduce some method of querying past events.