fix: Don't call into user toString
This implementation is naive, but much better than the previous version.
Instead of calling `toString` on user classes, return the class name and
object ID. Otherwise, if the class is part of the 'java.*' package
space, allow `toString` to occur. This should dramatically reduce the
amount of side effects we incur at runtime.