feat: Add session recording checkpoint
Stream all events to a temp file; removing the Memory recorder.

Recording is returned as an object, which can be deleted, or renamed,
or read into a Writer.

ActiveSessionException is sometimes meaningful, but usually it's just
like a RuntimeException. So, a lot of this behavior has been removed.
The main explicit check that remains is to see if a recording is or is
not in progress when certain actions are performed. Other types of
unexpected errors are not client-caused and therefore they are just
runtime exceptions and will probably result in a 500 error to the
client.

Because events are written into a temp file, the name of the AppMap file
(if any), is not needed until the recording is stopped.