fix: avoid a javassist issue
Add the @ArgumentArray annotation to the exception-handling hooks for
java.sql methods. The arguments aren't used, so passing an array of
Objects is fine.

Additionally, having a bunch of overloaded methods that could
appear in catch blocks seemed to confuse javassist. They caused it to
generate code that failed bytecode verification in some JVMs.