fix: don't depend on spring-web
Remove the dependency on org.springframework:spring-web, it should have
been removed as part of the changes to implement request recording.

Its inclusion prevents hooking of
SpringServletContainerInitializer.onStartup, which keeps request
recording doesn't work for a Spring app deployed as a WAR. So, these
changes also add a test for such an app.