fix: make sure events are flushed correctly
Separate the last event written to the global recording from the last
event written to the current thread's request recording.  Previously,
because there was only a single event, calling flush on one thread would
write the last event for other threads. This led to events winding up in
the wrong AppMaps.