fix: have safeClassForName require a ClassLoader
When resolving a class name, force the caller to choose a ClassLoader to
use for the resolution. Most of the time, this should be the ClassLoader
used to load a particular class. Occasionally, there isn't such a class,
so the thread's current class loader should be used instead.

Also, add ClassUtil.getMethodLocation as a way to determine the source
location of a method based on its signature.

Also, refactor the code that gets the value of an Enum field by name and
creates an EnumSet of that value.