fix: avoid Gradle class filtering
When running tasks, Gradle uses a class loader to filter out classes
that aren't on the implementation class path. This means that the agent
classes will get filtered out (unless the build configuration puts the
agent jar on the class path, which would be weird).

Avoid this filtering by adding "com.appland" as one of the packages that
should be allowed.