test: fix tests so they'll run on Windows
Make various tweaks to allow tests to run in a Git Bash environment on a
GitHub build runner. This includes disabling a couple of tests that have
proved challenging to port. They run fine on other platforms, though,
which should be sufficient to show that the functionality works.

These changes also converts the files containing those tests to JUnit 5.