fix: only add the agent jar if it can be located
The CodeSource and location of the agent jar's ProtectionDomain can both
be null (though the latter was only documented post-1.8). Only try to
add the jar if both are available.

Also, when the location URL is not null, convert it to a URI instead of
a Path. This works on all platforms.