feat: record calls made through a Proxy
When application code wants to call a method via a proxy created by
java.lang.reflect.Proxy, the proxy uses the invoke method of an
InvocationHandler to handle the request. These changes add a hook for
InvocationHandler.invoke which examines the Method argument. If it
matches the current config, it will be recorded in the AppMap just like
a regular method (i.e. one that's a member of a class).