fix: go back to hooking org.junit.jupiter.api.Test
JUnit 5 test execution listeners get notified before @BeforeEach
methods, and after @AfterEach methods. This causes some problems (e.g.
a temporary recording file can be deleted before it gets moved to its
final destination), but is also not what we want. By design, none of the
other agents capture test setup and teardown, so we shouldn't either.

These changes go back to just hooking JUnit 5's Test class. The change
to have instrumentation catch Throwable instead of exception was a
prerequisite for these changes, because AssertionError (raised when an
assertion fails) is an Error, not an Exception.

Also, change .metadata.framework to .metadata.frameworks so it matches
the spec.