build: bump shadow to 7.1.2
The previous version of shadow was failing when it encountered a Java 17
class file. None of our dependencies require Java 17, and it wasn't
immediately obvious exactly which class file was causing the failure.
Upgrading to 7.1.2 resolves the issue, though, and still results in an
agent that will run in Java 8+.