fix: manage ClassPool instances per-thread
To work in all environments, we need to manage instances of ClassPool
properly. In particular, we can't ever use ClassPool.getDefault -- it
initializes a ClassPool the first time it's used, which may or may not
reference the correct ClassLoader hierarchy.

Instead, we'll ensure each thread has its own ClassPool, linked to the
correct ClassLoaders.

Additionally, these changes fix a couple of problems exposed by
petclinic's recent upgrade to Spring Boot 3.2:

  * ReflectiveType and DynamicReflectiveType now handle
    InvocationTargetException correctly, rethrowing the target
    exception.

  * the charset of a JSON response is now UTF-8, so the number of bytes
    returned from /showavailable is less. Rather than make the test
    sensitive to the charset, we just check to make sure the count is
    greater than 0.