feat: create a log file by default
Update the logging configuration so a log file is created by default.
Inform the user of the file's location, and write some basic information
to it on startup and shutdown.

Also, fix a problem that caused messages to get dropped when exiting,
and fix a couple of minor issues in test scripts.