fix: make sure getters and setters are ignored
Fix the regex used to match the return type for getters and setters.

Making this fix identified another problem: the Spark/Jetty integration
depends on being able to hook a method named "setHandler". Because
getters and setters were being excluded before deciding to hook a
method, fixing the regex broke the integration. These changes move the
exclusion logic so it's only applied to methods matched by the config
(i.e. application code).

Finally, these changes add some tests for getter/setter exclusion, and
fix a test that was broken because it depended on a getter being
instrumented.