fix: make sure TestNG tests are captured correctly
Fix the hooks for TestNG tests, look for the TestNG @Test annotation
when scanning the stack for a test method.

These changes also include an update to the hooks for InvocationHandler.
Gradle's TestNG runner has a class that implements InvocationHandler,
but also has other overloads for "invoke". We need to make sure to only
hook the one we care about.