fix: don't try to add event on test failure
Previously, the test framework integrations all tried to add a "return"
event when an exception was thrown from a test method. This isn't right,
though, as they didn't add a "call" before entering the test method.
These changes remove the extra "return" event.

(Most of the time, this extraneous "return" event was ignored in
Recorder.add, which seems appropriate. Occasionally, for reasons I can't
determine, it was not. When it was not, fields in the event would get
cleared out, and lead to
https://github.com/getappmap/appmap-java/issues/215.)