fix: skip some JDK packages
Add a couple more top-level packages that are always skipped when
instrumenting. It's not clear to me that we'll never want to instrument
classes in these packages, though, so add a property to configure them.

Also, add a property to disable JGit when debugging.  It starts a thread
that loads classes, and there's no way to disable it. If you're trying
to debug something else when it starts, you suddenly wind up in that
thread.