fix: handle a framework classloader that filters
There are situations where the classloader hierarchy created by a
framework exerts tight control over the classes it will load. This can
lead to a ClassNotFoundException when code instrumented by the agent
tries to reference agent classes.

For example, prior to these changes, trying to instrument an app built
on the Play Framework would throw a ClassNotFoundException trying to use
EventTemplateRegistry.

Now, at runtime, we add a class to the boot class loader that provides
the instrumented code with a way to create "call" and "return" events.