refactor: change Recorder to an Enum singleton
Switch Recorder from a class to an enum, add the INSTANCE member. This
simplifies the code a little, and follows a pretty common pattern.

This also required switching the Mockito mocker to the inline mocker.
This mocker became the default in Mockito 5, but we can't just upgrade
because 5 requires Java 11.