fix: resolve output directory relative to config
The appmap_dir specified in a config file should be resolved relative to
the location of the config file.

For example, if appmap.yml is in the root of project, contains
"appmap_dir: tmp/appmap", and is discovered for tests in a submodule
run by maven, the AppMaps will be in "$project_root/tmp/appmap". Prior
to these changes, the recordings were placed in
"$project_root/$submodule_dir/tmp/appmap".