build: Download petclinic archive instead of full clone
To speed up CI, download a tarball of a specific commit hash
instead of performing a full git clone. This is particularly
useful for pinning spring-petclinic to a version compatible
with Java 17.

After extracting the tarball, a new git repository is
initialized to simulate a checkout (which the tests expect).

Refactor the script to handle different Java versions and use
the appropriate petclinic version for each.

This commit also updates the appmap-labels.yml file to
correctly identify the logging classes used in newer versions
of the project.